pyproject.toml
[tool.poetry]
name = "stack3"
version = "0.1.0"
description = ""
authors = ["ng14th <thnguyen.github@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
motor = "^3.1.1"
fastapi = "^0.88.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
umongo = "^3.1.0"
aioredis = "^2.0.1"
ujson = "^5.6.0"
celery = "^5.2.7"
redis = "^4.4.0"
fastapi-utils = "^0.2.1"
flower = "^1.2.0"
pyjwt = "^2.6.0"
bcrypt = "^4.0.1"
nats-py = "^2.2.0"
nats-python = "^0.8.0"
aiormq = "^6.6.4"
kombu = "^5.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
